refactor: move Feed to utils.py
