+++ main.py
@@ -1,48 +1,12 @@
 import datetime
 import os
-import pprint
-from random import randint
-from typing import List, Type
+from typing import List
 
 from feedgen.feed import FeedGenerator
 
 from tg_api import TGPostDataclass, TGApiChannel
-from utils import ContentItem, ApiClass, shortened_text
+from utils import shortened_text, Feed
 from yt_api import YTVideoDataclass, YTApiChannel
-
-
-class Feed:
-    ContentItemClass = ContentItem
-    api_class: Type[ApiClass] = ApiClass
-
-    def __init__(self, url: str):
-        self.url = url
-        self.api_object = self.api_class(url)
-
-    def fetch_all(self, after_date: datetime.date = None) -> List[ContentItem]:
-        """
-        Base function to get new updates from given feed.
-        Must be overridden by every Sub-class.
-        :return: List[ContentItem]
-        """
-        if after_date:
-            if self.api_class.SUPPORT_FILTER_BY_DATE:
-                self.api_object.published_after_param = after_date
-            else:
-                result = list()
-                try:
-                    for i in iter(self.api_object):
-                        if i.pub_date > after_date:
-                            result.append(i)
-                        else:
-                            raise StopIteration
-                except StopIteration:
-                    pprint.pprint(result)
-                    return result
-
-        result = list(self.api_object)  # Invokes generator with http requests
-        pprint.pprint(result)
-        return result
 
 
 class YTFeed(Feed):
